# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DiceGame",
        "description": "URI's List for DiceGameApi",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/v1/players/{id}/games": {
            "get": {
                "tags": [
                    "Player"
                ],
                "summary": "Get games history of a player",
                "operationId": "gamesHistory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the player",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Player John Doe games history:"
                                        },
                                        "games_played": {
                                            "type": "integer",
                                            "example": "5"
                                        },
                                        "success_rate": {
                                            "type": "string",
                                            "example": "75.00 %"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Game"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/players/{id}/games-history": {
            "delete": {
                "tags": [
                    "Player"
                ],
                "summary": "Delete game history for a specific player",
                "description": "Delete game history for a specific player.",
                "operationId": "deleteHistory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the player",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game history deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Game history deleted for user: John Doe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Game": {
                "title": "Game",
                "description": "Game schema",
                "required": [
                    "user_id",
                    "dice1_value",
                    "dice2_value",
                    "sum",
                    "result"
                ],
                "properties": {
                    "user_id": {
                        "description": "ID of the user associated with the game",
                        "type": "integer"
                    },
                    "dice1_value": {
                        "description": "Value of the first dice rolled in the game",
                        "type": "integer"
                    },
                    "dice2_value": {
                        "description": "Value of the second dice rolled in the game",
                        "type": "integer"
                    },
                    "sum": {
                        "description": "Sum of the values of the two dice",
                        "type": "integer"
                    },
                    "result": {
                        "description": "Result of the game",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}